server: Switch management policy to activity score
This changes the policy for closing channels from an absolute policy
(close all channels that don't meet the policy conditions) to a relative
policy (based on activity scoring for each channel).

The policy enforcement only kicks off after the on-chain wallet balance
drops below a new config parameter. The intended result is that channels
will not be closed unless the liquidity provider requires coins to fund
new channels, preferring to close channels with a combination of low
activity during its lifetime.

This is a more useful policy in the current state of Decred's LN due to
low rates of use.